api/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
sqlparse==0.4.3
django-cors-headers
mysqlclient==2.1.1
Pillow==9.4.0
djangorestframework
djangorestframework-simplejwt
cloudinary
drf-writable-nested